refactor : 메서드 역할을 cars객체에게 부여
