+++ src/main/java/racingcar/domain/Cars.java
@@ -1,7 +1,9 @@
 package racingcar.domain;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
+import java.util.stream.Collectors;
 
 public class Cars {
     private final List<Car> cars;
@@ -13,6 +15,22 @@ public class Cars {
 
     public void addCars(String carName) {
         cars.add(new Car(carName));
+    }
+
+    public List<Car> selectWinners() {
+        int winnerScore = selectWinnerScore(this.cars);
+
+        return this.cars.stream()
+                .filter(cars -> cars.isWinner(winnerScore))
+                .map(cars -> new Car(cars.getName()))
+                .collect(Collectors.toList());
+    }
+
+    private int selectWinnerScore(List<Car> carInfo) {
+        return carInfo.stream()
+                .max(Comparator.comparingInt(Car::getLocation))
+                .get()
+                .getLocation();
     }
 
     public void setCycle(int cycle) {
